Edit to first sentence
Edited the sentence "A game kinda like space invaders, but the invaders evolve with genetic algorithm" to "InvaderZ is like the game Space Invaders, except InvaderZ uses the genetic algorithm to mutate the invaders as you play." The sentence makes it more clear to readers that InvaderZ uses the genetic algorithm to change the invaders as you play. The phrase "invaders evolve with genetic algorithm" could mean that the genetic algorithm changes as the invaders change, not the genetic algorithm causing the change in the invaders.